scenario decorator importing

+++ blender/addons/io_scene_foundry/managed_blam/decorator_set.py
@@ -35,7 +35,7 @@ class DecoratorSetTag(Tag):
                     
         return type_names
         
-    def to_blender(self, collection, get_material=False, always_extract_bitmaps=False):
+    def to_blender(self, collection, get_material=False, always_extract_bitmaps=False, single_type=None, highest_lod_only=False, only_single_type=False):
         base = self.get_path_str(self.reference_base.Path, True)
         lod2 = self.get_path_str(self.reference_lod2.Path, True)
         lod3 = self.get_path_str(self.reference_lod3.Path, True)
@@ -43,54 +43,75 @@ class DecoratorSetTag(Tag):
         texture = self.get_path_str(self.reference_texture.Path, True)
         
         all_obs = []
+        found_highest_lod = False
+        single_type_index = None
+        
+        types = self.get_type_names()
+        
+        if single_type:
+            
+            for idx, t in enumerate(types):
+                if t == single_type:
+                    single_type_index = idx
+                    break
+                
+        if single_type_index is None and only_single_type and types:
+            single_type_index = 0
         
         if base and Path(base).exists():
             base_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_high")
             collection.children.link(base_collection)
             with RenderModelTag(path=base) as base_render:
-                base_obs = base_render.to_blend_objects(collection, True, False, base_collection, no_armature=True)
+                base_obs = base_render.to_blend_objects(collection, True, False, base_collection, no_armature=True, specific_io_index=single_type_index)
                 for ob in base_obs:
                     ob.nwo.decorator_lod = 'high'
                     utils.unlink(ob)
                     base_collection.objects.link(ob)
                     
                 all_obs.extend(base_obs)
+                found_highest_lod = True
+        
+        if not highest_lod_only or not found_highest_lod:
+            if lod2 and Path(lod2).exists():
+                lod2_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_medium")
+                collection.children.link(lod2_collection)
+                with RenderModelTag(path=lod2) as lod2_render:
+                    lod2_obs = lod2_render.to_blend_objects(collection, True, False, lod2_collection, no_armature=True, specific_io_index=single_type_index)
+                    for ob in lod2_obs:
+                        ob.nwo.decorator_lod = 'medium'
+                        utils.unlink(ob)
+                        lod2_collection.objects.link(ob)
+                        
+                    all_obs.extend(lod2_obs)
+                    found_highest_lod = True
+        
+        if not highest_lod_only or not found_highest_lod:
+            if lod3 and Path(lod3).exists():
+                lod3_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_low")
+                collection.children.link(lod3_collection)
+                with RenderModelTag(path=lod3) as lod3_render:
+                    lod3_obs = lod3_render.to_blend_objects(collection, True, False, lod3_collection, no_armature=True, specific_io_index=single_type_index)
+                    for ob in lod3_obs:
+                        ob.nwo.decorator_lod = 'low'
+                        utils.unlink(ob)
+                        lod3_collection.objects.link(ob)
                     
-        if lod2 and Path(lod2).exists():
-            lod2_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_medium")
-            collection.children.link(lod2_collection)
-            with RenderModelTag(path=lod2) as lod2_render:
-                lod2_obs = lod2_render.to_blend_objects(collection, True, False, lod2_collection, no_armature=True)
-                for ob in lod2_obs:
-                    ob.nwo.decorator_lod = 'medium'
-                    utils.unlink(ob)
-                    lod2_collection.objects.link(ob)
-                    
-                all_obs.extend(lod2_obs)
-                
-        if lod3 and Path(lod3).exists():
-            lod3_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_low")
-            collection.children.link(lod3_collection)
-            with RenderModelTag(path=lod3) as lod3_render:
-                lod3_obs = lod3_render.to_blend_objects(collection, True, False, lod3_collection, no_armature=True)
-                for ob in lod3_obs:
-                    ob.nwo.decorator_lod = 'low'
-                    utils.unlink(ob)
-                    lod3_collection.objects.link(ob)
-                
-                all_obs.extend(lod3_obs)
-                
-        if lod4 and Path(lod4).exists():
-            lod4_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_ver_low")
-            collection.children.link(lod4_collection)
-            with RenderModelTag(path=lod4) as lod4_render:
-                lod4_obs = lod4_render.to_blend_objects(collection, True, False, lod4_collection, no_armature=True)
-                for ob in lod4_obs:
-                    ob.nwo.decorator_lod = 'very_low'
-                    utils.unlink(ob)
-                    lod4_collection.objects.link(ob)
-                    
-                all_obs.extend(lod4_obs)
+                    all_obs.extend(lod3_obs)
+                    found_highest_lod = True
+        
+        if not highest_lod_only or not found_highest_lod:
+            if lod4 and Path(lod4).exists():
+                lod4_collection = bpy.data.collections.new(f"{self.tag_path.ShortName}_ver_low")
+                collection.children.link(lod4_collection)
+                with RenderModelTag(path=lod4) as lod4_render:
+                    lod4_obs = lod4_render.to_blend_objects(collection, True, False, lod4_collection, no_armature=True, specific_io_index=single_type_index)
+                    for ob in lod4_obs:
+                        ob.nwo.decorator_lod = 'very_low'
+                        utils.unlink(ob)
+                        lod4_collection.objects.link(ob)
+                        
+                    all_obs.extend(lod4_obs)
+                    found_highest_lod = True
         
         if texture and Path(texture).exists():
